Removed warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@10873 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 185b8d03dcc9fbf4f0b85ce667241dc316f5c260

+++ costmodel/de.uniol.inf.is.odysseus.costmodel.operator.datasrc/src/de/uniol/inf/is/odysseus/costmodel/operator/datasrc/impl/histogram/ScottsRule.java
@@ -22,7 +22,7 @@ public class ScottsRule implements IIntervalCountEstimator {
 		}
 		
 		// erwartungswert
-		double e = sum / (double)data.size();
+		double e = sum / data.size();
 		
 		// variances
 		double sum2 = 0.0;
